requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
pycryptodome==3.20.0
qrcode==7.4.2
reportlab==4.1.0
pyzbar==0.1.9
pdf2image==1.17.0